Add MacOS support to STP_USE_LIB
On MacOS, libraries have a different suffix and the filesystem is structured differently.
This commit adds the relevant bits and pieces for the option STP_USE_LIB to work on MacOS
in principle.

Also fixes the issue with STP_STUB (which is used with STP_USE_LIB=disable), where the wrong
library file is copied, resulting in a linking error down the line.