Improve readability in stp/Makefile
The logic for STP_USE_LIB has been indented to improve the readability and
make the different cases more obvious.

Also, minor fixes. `find` ends with NUL byte, so replace that with newline.
The BSC linker needs at least a symlink named libstp.so
The SNAME is completely independent from the library name, so just choose a fitting one