Introduce placeholder option to STP
Using STP_USE_LIB=placeholder means that the stub is compiled with a generic name and
copied to the installation, but is supposed to be replaced at a later date.

To make the intentions clear, the library is renamed to *_placeholder with a symlink
created with the original name pointing to the library.