Fixes to typecheck of struct update/literal/pattern
This addresses GitHub issue 628, by fixing the desugaring of struct
updates during typecheck.  It mostly fixes issues around qualified
field names, which is only allowed in BH, but it also fixes the
position of the fields in the new expression, which might improve
position information for errors in BSV.

Test cases are added for both qualified and unqualified field names in
struct update, construction, and pattern matching.  Further tests are
added for incorrect qualifiers and duplicate field names with the same
base name (such as qualified and unqualified).

The typecheck of CStruct had been fixed in 2006 to account for
qualifiers (in most situations), but similar code in the typecheck of
CStructUpd that needed the same fix was overlooked, and typecheck of
CPstruct was even further behind.  These are all updated, with
additional comments and some improved variable names.