Workaround a PATH/ghc issue on macOS 11 VM
Homebrew has stopped providing pre-built packages for macOS 11, so
packages are built from source, which requires installing any tools
needed to build the source.  Thus, ghc and cabal are now being
installed in /usr/local/bin, which is earlier in the PATH than the
ghcup bin directory, so the wrong GHC is found.  This workaround
unlinks the specific ghc and cabal-install packages.  Another option
would be to put ghcup earlier in the PATH.