Reorder ASchedule
so that the order of code matches the order of the compuation.  The
initial conflict graphs are computed and then edges are added to those
graphs, but the code to compute the edges was appearing first.
This reorder will allow the addition of code for new edges that needs
to consult the existing graph first (to avoid unnecessary work).
A consequence of this reordering, though, is that the forcing of the
graphs is delayed until all three are constructed.