Bump GHC version to 9.4.7, fix HLS at 2.5.0.0
Because we weren't specifying the HLS version for ghcup to install,
but were specifying the GHC version, the GitHub CI suddenly started
failing when a new HLS version became the default and which didn't
support the GHC version we were using because a newer patch version
became available.  To avoid that, we fix the HLS version, so that we
can update the GHC and HLS versions together, on our schedule.