Add GHC version to Bluetcl, for use in testsuite
The Bluetcl 'version' command now has a 'ghc' subcommand that will
print the version of GHC that the Bluespec tools were compiled with.
The subcommand 'bsc' was also added, which has the same behavior as
calling 'version' with no arguments: prints the Bluespec tools version.
The auotmated printing of help messages needed to be updated to work
for commands with optional subcommands like this.

The testsuite now has a variable 'ghc_version' which can be consulted.
The tests for bug 1490 are updated to use this, because BSC compiled
with GHC 9.8.1 has regressions that need to be handled differently.

The regexp for assigning 'bsc_version' in the testsuite was broken and
was assigning it multiple lines of text (which caused confusion when
printed in the log).  Instead of fixing the regexp, the getting of the
version is changed to use Bluetcl, to print just the version info.