Make macOS CI more robust to changes in the SystemC library
The GitHub VMs for macOS 12 and 13 are now installing SystemC compiled
with a newer C++ standard than C++11.  This causes the testsuite to
fail to link, because we have hardcoded the flag '-std=c++11'.
In fact, the macOS 12 VM seems to sometimes install a library using
C++11 and sometimes using C++17.  To account for this, identify the
flag to use by looking at the symbols in the library.