Typecheck: Do more substitution on reduced preds
Add a test case that fails to typecheck without the substitution.
It's unclear if this is the most efficient fix -- for example, if the
substitution is just covering for an issue earlier where the
predicates are created -- but this doesn't seem to increase the run
time of the testsuite, so it's probably ok.