Update test for Bluespec Inc bug 1490
The heap overflow in GHC 9.8 (actually starting with 9.6) is correct.
Earlier GHC 9.4 can be seen to use the same amount of memory, but is
not correctly handling the RTS -M flag.  If possible, the exp file
should be updated to more directly test what the concern is -- whether
that's a trace of the BSC elaboration or GHC's reporting of the live
bytes (which is smaller than the actual heap usage, due to
fragmentation) or something else.