GitHub CI: Update brew/apt-get before install
The macOS VMs have recently failed to install mactex, because the
download URL is not found.  This is probably because "brew update" is
not being called, to update the formulae.  On Ubuntu, where apt-get is
used to install, in most cases "apt-get update" was being called, but
there was one place (Toooba test) that was missing.