GitHub CI: Install pyyaml via brew not pip, for macOS 14+
Installing with pip3 on the macos-14 VM gives an error that the python
environment is externally managed.  It suggests using brew to install
packages, or else to create a virtual environment (with "venv"), or
use pipx to manage a virtual environment for you.  Unfortunately, the
brew install doesn't seem to be visible on the older VMs.  It may be
possible to use the same "venv" commands across VMs, but it's easy
enough to check the macOS version and use brew or pip3 to install as
appropriate.