Testsuite: Add "which_mach" and use in div-by-zero tests
The divmod tests fail on arm64, because the tests are expecting SIGFPE
error to be raised, but ARM does not.  This may need to be addressed in
the code generation for Bluesim, but in the meantime, update the test
to expect no error when the MACHTYPE is arm64.

A new "which_mach" procedure in the testsuite allows querying MACHTYPE.

Found an old test that was calling "uname -m" to get the machtype, and
test for x86_64.  This could be replaced with a call to "which_mach",
but the use was to workaround a bug in iverilog 0.8.3, which is older
than we need to support, so instead the code is just removed.