GitHub CI: Install pyyaml via apt not pip, for ubuntu 24.04+
Installing with pip3 on the ubuntu-24.04 VM gives an error that the
python environment is externally managed.  It suggests using "apt" to
install packages, or else to create a virtual environment (with "venv"),
or use pipx to manage a virtual environment for you.  This issue was
encountered for macos-14 and addressed in an earlier commit.  We apply
a similar patch here.  It's possible that the same "apt" command can be
used for all Ubuntu versions; but we check the Ubuntu version and only
use "apt" for 24+.