Bluesim: Fix function call arguments in debug function
The EventQueue::print() function was calling bk_clock_name() with the
wrong number of arguments, but fortunately print() is not used anywhere,
as it only exists for use when debugging.  The issue is resolved by
adding an additional parameter to print().  However, since so many
functions in the EventQueue class take the tSimStateHdl as an argument,
it might make sense for the EventQueue constructor to take that value
and store it.