solver: remove stp support entirely
The version of STP vendored in the repo is over 10 years old, and Yices2 is the
default. Even though it has other complications (like GPLv3), it's pointless to
keep both of these around when the real path forward is to most likely support
arbitrary solvers via smtlib.

Delete this incredibly old snapshot and all references to it, and enjoy the
improved compile times, and just rely on Yices for now.

Signed-off-by: Austin Seipp <aseipp@pobox.com>