src: move `vendor/htcl/` under `comp/`
Similarly to Parsec, this code is practically only used in one spot, so it
might as well be maintained there. I don't even know what codebase HTcl is
a 'vendoring' *of* and it's dated to over 15 years ago at minimum, so it's
probably lost to history now... Just put it all together for clarity and ease
of development.

Signed-off-by: Austin Seipp <aseipp@pobox.com>