comp: remove most imports of `Data.Generics`
99% of the uses of the `Data.Generics` import from `syb` was to qualify the
names `Data` and `Typeable`, but these have long since been available directly
in `base` along with `DeriveDataTypeable` for many years now. The `syb` exports
are in fact just re-exports from base. Migrate most of the code to just use
`Data.Data` directly instead.

At the same time, this touches up some of the import lists to be a little more
consistent, but is otherwise functionally identical.

Signed-off-by: Austin Seipp <aseipp@pobox.com>