Testsuite: Support a list of possible compare files
The 'compare_file' procedure could be updated to check whether the
'expected' argument is a list, but in the meantime, this adds a new
'compare_file_list' procedure that takes a list argument and passes
if any one in the list matches.

This is used in one test where BSC built with GHC 9.12.1 has _tcdict
defs in a different order.  The test now compares against both
possible outputs.