Improve signal names generated from real literals
If a real literal was small enough to use E notation with a negative
exponent, that negative sign would appear in signal names derived from
that expression, which is illegal for both Verilog and Bluesim.  We
were eliminating decimal points from numbers, but failing to remove
negative signs.  Instead of removing them, however, we now replace
them (with "d" and "neg") so that the value in the name is not
misinterpreted.