Find tcl8 on Fedora-based systems
With the release of tcl 9, tcl8 commands get installed with a suffix,
e.g. 'tclsh8'. platform.sh rejects tcl 9 as unsupported, and fails to
find tcl 8 when installed with a suffix.

Since 'tclsh8' more specifically identifies a tcl version that's
definitely supported, look for that first, and then try the generic
'tclsh' as a fallback.

Signed-off-by: David Anderson <dave@natulte.net>