Silence "unused result" warning in bluesim probe module
In modern glibc, asprintf's return value is annotated as "must use",
because it can fail with -1 in OOM conditions.

This is not a correctness issue on modern OSes, since the surrounding
code will at worst lead to a null pointer deref and crash in this
condition. Further, most modern OSes allocate virtual memory
optimistically, and crash later on failure to fault in rather than
returning -1 to an allocation call. So, this change's main purpose
is to declutter bluesim compilation logs.

Signed-off-by: David Anderson <dave@natulte.net>