INSTALL.md: reorganize build requirements, add instructions for Fedora
The reorganized requirements section opens with a full list of required and
optional dependencies in plain prose. Following that, separate subsections
provide handy cut-and-paste commands to run on Debian, Ubuntu and Fedora in
order to install those dependencies, as well as instructions on using
Cabal directly to manage GHC libraries.

The sections describing the SMT solvers and how to disable them were also
restructured slightly, to heal them from the removal of interleaved package
install instructions.

Signed-off-by: David Anderson <dave@natulte.net>