Centralize Tcl version checking into 'platform.{sh,mk}'
Add 'tclversion' subcommand to 'platform.sh' and provide it in
'platform.mk' as 'TCL_VERSION'.  Move the check of the Tcl version out
of the HTcl makefile and into 'platform.mk' where it can be shared
across the BSC build as necessary.  This check also constructs
'TCL_DEFS' as the list of macro definitions that BSC code requires for
that version of Tcl (currently 'TCL85' and 'TCL9').  These can be
preceded by '-D' or '-optc-D' and passed to the C compiler or GHC,
where needed.  One use is in the HTcl Makefile and another will be
added later to 'src/comp/' Makefile, to support conditional code in
the 'Htcl.hs' (which is compiled by that Makefile, not the one in the
HTcl vendor directory).