Support Tcl 9 and when Tcl functions are macros
For the GHC foreign import of Tcl/HTcl, use 'capi' instead of 'ccall'
where the function may possibly be defined via macro and not a real
function.  This helps even with Tcl 8 on systems like Windows where
the interface is defined with macros.  On Tcl 9, even in Linux some
functions are macros.  Newer versions of GHC are more strict about
matching the 'const' type when using 'capi', so use 'ConstPtr' for
function arguments that are 'const' when the GHC base version is
4.18.0 or greater.

Tcl 9 no longer limits indices to 32 bits for 64-bit architectures
(which we assume).  Therefore, we replaced hardcoded uses of 'CInt'
with 'CTclSize', which is defined as 'CInt' or 'CLLong' depending on
whether the Tcl version is 8 or 9.