INSTALL.md: reorganize and add instructions for Fedora and macOS
The reorganized requirements section opens with a full list of required and
optional dependencies in plain prose. Following that, separate subsections
provide handy cut-and-paste commands to run on Debian/Ubuntu, Fedora, and
macOS, in order to install those dependencies, as well as instructions on
using Cabal directly to manage GHC libraries.

The sections describing the SMT solvers and how to disable them were also
restructured slightly, to heal them from the removal of interleaved package
install instructions.

Tcl requirements have been updated.  And GMP has been included.

This also replaces the indent-plus-dollar-sign convention for
commands with bash blockquotes, to be consistent throughout.

---------

Signed-off-by: David Anderson <dave@natulte.net>
Co-authored-by: Julie Schwartz <quark@bluespec.com>