Teach bluetcl Bluespec Haskell syntax.
The syntax MutableVar is exposed to bluetcl via new commands.
Every place where bluetcl used a pv printing function it now uses a pf function.
Output syntax is directly adjusted in a few places.

When bluetcl parses types or constructors, it now supports Bluespec Haskell and
Bluespec SystemVerilog syntax. Parses using the current mode are preferred,
but if parsing fails in one syntax the parser is permissive and tries
to parse with the other syntax.