Move primSeq calls to minimize retention in the ActionValue Monad
With the new implementation, all of the fields of the previous AVStruct
will be evaluated before the ActionWorld computation is started for the
next AVStruct, which should hopefully minimize the retained references
to previous computations.

Update expected file for moved definition in output.

Update id of DEF_v__h<n> in expected file.
I'm presuming the reason this isn't a regexp is because we want to make sure
the identifier matches between the two searches.