GitHub CI: Fix the checkout-same-branch feature
For one, if the user doesn't have a fork of the repo, the script would
fail, rather than checkout from the default location.

Two, the GITHUB_ACTOR field is not always the owner of the head repo
for a PR.  If someone else adds to the branch (including rebasing),
that person is the actor.  Unfortunately, while this information is
available in the workflow context, it is not available in any of the
automatic environment variables, so we have to define our own in each
place that we call the script.  The script expects the variable
HEAD_OWNER to be defined for pull requests.